# [AVF] Optimize MediaSampleAVF timing access