#### f5f2798ff872749c51d48b4609da98e1735622a5
<pre>
[AVF] Optimize MediaSampleAVF timing access
<a href="https://bugs.webkit.org/show_bug.cgi?id=244322">https://bugs.webkit.org/show_bug.cgi?id=244322</a>
&lt;rdar://99119415&gt;

Reviewed by Jer Noble.

MediaSampleAVF calls directly into CoreMedia in order to fetch a media
timing which is then converted into a MediaTime object. Considering how
frequently these timings are accessed, we can optimize the access speed
by holding our own MediaTime object and caching the underlying timings.

Benchmark (ran 4150 times):
```
    MediaTime x;
    {
        TimingScope(&quot;sample&quot;);
        for (int i = 0; i &lt; 10000; i++)
            x += sample-&gt;presentationTime();
    }
```
before: 4150 calls, mean duration: 0.000022ms, total duration: 0.091249ms, max duration 0.000792ms
after: 4150 calls, mean duration: 0.000015ms, total duration: 0.062080ms, max duration 0.000333ms

* Source/WebCore/platform/graphics/avfoundation/objc/MediaSampleAVFObjC.h:
* Source/WebCore/platform/graphics/avfoundation/objc/MediaSampleAVFObjC.mm:
(WebCore::MediaSampleAVFObjC::MediaSampleAVFObjC):
(WebCore::MediaSampleAVFObjC::initializeTimes):
(WebCore::MediaSampleAVFObjC::presentationTime const):
(WebCore::MediaSampleAVFObjC::decodeTime const):
(WebCore::MediaSampleAVFObjC::duration const):
(WebCore::MediaSampleAVFObjC::offsetTimestampsBy):
(WebCore::MediaSampleAVFObjC::setTimestamps):

Canonical link: <a href="https://commits.webkit.org/253796@main">https://commits.webkit.org/253796@main</a>
</pre>